backend/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
requests==2.32.3
python-dotenv==1.0.1
openai==1.51.0
vonage==3.14.0
sendgrid==6.11.0
redis==5.1.1
bcrypt==4.2.0
marshmallow==3.22.0
celery==5.4.0
gunicorn==23.0.0
psycopg2-binary==2.9.9
